+++ scripts/macros/classFeatures/cleric/deathDomain/reaper.js
@@ -37,6 +37,8 @@ export async function reaper({speaker, actor, token, character, item, args, scop
         'configureDialog': false,
         'versatile': false,
         'consumeResource': false,
+        'consumeQuantity': false,
+        'consumeUsage': false,
         'consumeSlot': false,
         'workflowOptions': {
             'autoRollDamage': 'always',