+++ scripts/macros/classFeatures/fighter/battleMaster/maneuvers.js
@@ -86,6 +86,8 @@ async function goadingAttack({speaker, actor, token, character, item, args, scop
             'configureDialog': false,
             'versatile': false,
             'consumeResource': false,
+            'consumeQuantity': false,
+            'consumeUsage': false,
             'consumeSlot': false,
             'workflowOptions': {
                 'autoRollDamage': 'always',