+++ scripts/macros/mechanics/cleave/cleave.js
@@ -79,6 +79,8 @@ export async function cleave(workflow) {
         'configureDialog': false,
         'versatile': false,
         'consumeResource': false,
+        'consumeQuantity': false,
+        'consumeUsage': false,
         'consumeSlot': false,
         'workflowOptions': {
             'autoRollDamage': 'always',