+++ scripts/macros/monsterFeatures/bulette/deadlyLeap.js
@@ -55,6 +55,8 @@ export async function deadlyLeap({speaker, actor, token, character, item, args, 
         'configureDialog': false,
         'versatile': false,
         'consumeResource': false,
+        'consumeQuantity': false,
+        'consumeUsage': false,
         'consumeSlot': false,
         'workflowOptions': {
             'autoRollDamage': 'always',