+++ scripts/macros/spells/sanctuary.js
@@ -34,6 +34,8 @@ export async function sanctuary(workflow) {
         'configureDialog': false,
         'versatile': false,
         'consumeResource': false,
+        'consumeQuantity': false,
+        'consumeUsage': false,
         'consumeSlot': false,
         'workflowOptions': {
             'autoRollDamage': 'always',